Extract file record loading partial
